Drop generated files from source control
These files are created by automake and autoconf and do not need to be
tracked in GIT.  Instead we supply a helper script, autogen.sh, that
can be used when building directly from GIT.  When creating a release,
with 'make dist' or 'make distcheck', the GNU build system creates a
tarball with the generated files included.

Reduce automake strictness from GNU (default) to "foreign".  We don't
need to follow the GNU coding standards in this project.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>