Simplify checks for struct members and missing std functions
The configure.ac AC_CHECK_STRUCT_FOR() can be replaced with AC_CHECK_MEMBER()
that incidentally are already used.  Dropping AC_CHECK_STRUCT_FOR() helps us
with the long-term plan to remove the AC_TRY_COMPILE() macro, troublesome for
cross-compiling use-cases.

The AC_CHECK_FUNC_PROTO() macro is also dropped.  Operating systems that do
not have strerror() and inet_aton() declared as prototypes in standard header
files are not worth the complexity of supporting.

Drop libreplace.a in favor of the built-in mechanism in automake to build and
link in replacements using $(LIBOBJS).  Needs a bit of subdir-objects trickery
but reduced overall complexity.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>