Drop GNU license files and rename COPYING -> LICENSE
There is no more GNU licensed code in the project, so we can now drop
these two files.  Automatically generated files by the GNU build system
do not require GPL/LGPL since they are not included in the resulting
binary files (c.f., gcc).

The filename COPYING is sort of claimed by GNU as the name of their
license.  So let's rename it to LICENSE, which is what most of any
other BSD-like licensed project use. (Principle of least surprise.)

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>